% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataRequest.r
\name{dataRequest}
\alias{dataRequest}
\title{Open csv requestData file}
\usage{
dataRequest(totalParams)
}
\arguments{
\item{totalParams}{dataframe}
}
\value{
requestData list
}
\description{
Open csv requestData file
}
\examples{
fileName <- "ULPParams3.csv"
filePath <- "//igsarmewfsapa/projects/QW Monitoring Team/GLRI beaches/Modeling/Model development 2012/ULP"
totalParams <- getParamsFromFile(filePath,fileName)
requestData <- dataRequest(totalParams)
}

