\name{getGLCFSData}
\alias{getGLCFSData}
\title{Get GLCFS Data from EnDDaT}
\usage{
  getGLCFSData(totalParams, baseReturn)
}
\arguments{
  \item{totalParams}{dataframe}

  \item{baseReturn}{string}
}
\value{
  GLCFSData dataframe
}
\description{
  Get GLCFS Data from EnDDaT
}
\examples{
\dontrun{getGLCFSData(totalParams,baseReturn)}
}

