\name{getNWISData}
\alias{getNWISData}
\title{Get NWIS Data from EnDDaT}
\usage{
  getNWISData(totalParams, baseReturn)
}
\arguments{
  \item{totalParams}{dataframe}

  \item{baseReturn}{string}
}
\value{
  NWISData dataframe
}
\description{
  Get NWIS Data from EnDDaT
}
\examples{
\dontrun{getNWISData(totalParams,baseReturn)}
}

