% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPRECIPData.r
\name{getPRECIPData}
\alias{getPRECIPData}
\title{Get PRECIP Data from EnDDaT}
\usage{
getPRECIPData(totalParams, baseReturn)
}
\arguments{
\item{totalParams}{dataframe}

\item{baseReturn}{string}
}
\value{
PRECIPData dataframe
}
\description{
Get PRECIP Data from EnDDaT
}
\examples{

\dontrun{getPRECIPData(totalParams,baseReturn)}
}

