% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glcfsCheatSheet.R
\name{glcfsCheatSheet}
\alias{glcfsCheatSheet}
\title{glcfsCheatSheet}
\usage{
glcfsCheatSheet()
}
\value{
baseURL string
}
\description{
glcfsCheatSheet shows abbrieviations need in parameter file for glcfs data
}
\examples{
baseReturn <- glcfsCheatSheet()
}

