% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGLCFSurl.r
\name{generateGLCFSurl}
\alias{generateGLCFSurl}
\title{Get GLCFS parts of URL for EnDDaT}
\usage{
generateGLCFSurl(xPoint, yPoint, zPoint, sourceNum, varName, colName)
}
\arguments{
\item{xPoint}{string GLCFS x index}

\item{yPoint}{string GLCFS y index}

\item{zPoint}{string GLCFS z index}

\item{sourceNum}{integer (0,1,2) defines which data source}

\item{varName}{string variable name}

\item{colName}{string column name}
}
\value{
GLCFScall string
}
\description{
Get GLCFS parts of URL for EnDDaT
}
\examples{
xPoint <- 130
yPoint <- 13
zPoint <- -1
sourceNum <- 0
varName <- 'vc'
colName <- 'Northward velocity at surface'

firstReturn <- generateGLCFSurl(xPoint, yPoint,zPoint,sourceNum,varName,colName)
}

