\name{getPRECIPData}
\alias{getPRECIPData}
\title{Get PRECIP Data from EnDDaT}
\usage{
  getPRECIPData(totalParams, baseReturn)
}
\arguments{
  \item{totalParams}{dataframe}

  \item{baseReturn}{string}
}
\value{
  PRECIPData dataframe
}
\description{
  Get PRECIP Data from EnDDaT
}
\examples{
\dontrun{getPRECIPData(totalParams,baseReturn)}
}

